+++ CHANGELOG.md
@@ -7,6 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+* `anyhow` feature flag which provide conversions from error types to `anyhow::Error`.
+
 
 ## [0.1.0] - 2023-11-23